backend/requirements.txt
anyio==4.13.0
asgiref==3.11.1
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
firebase_admin==7.4.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-core==2.5.1
google-cloud-firestore==2.27.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
msgpack==1.1.2
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
requests==2.33.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
